Unpack the initial channel only once

+++ modules/installer/cd-dvd/channel.nix
@@ -29,10 +29,14 @@ in
   # for nixos-install.
   boot.postBootCommands =
     ''
-      echo "unpacking the NixOS/Nixpkgs sources..."
-      mkdir -p /nix/var/nix/profiles/per-user/root
-      ${config.environment.nix}/bin/nix-env -p /nix/var/nix/profiles/per-user/root/channels -i ${channelSources} --quiet
-      mkdir -m 0700 -p /root/.nix-defexpr
-      ln -s /nix/var/nix/profiles/per-user/root/channels /root/.nix-defexpr/channels
+      if ! [ -e /var/lib/nixos/did-channel-init ]; then
+        echo "unpacking the NixOS/Nixpkgs sources..."
+        mkdir -p /nix/var/nix/profiles/per-user/root
+        ${config.environment.nix}/bin/nix-env -p /nix/var/nix/profiles/per-user/root/channels -i ${channelSources} --quiet
+        mkdir -m 0700 -p /root/.nix-defexpr
+        ln -s /nix/var/nix/profiles/per-user/root/channels /root/.nix-defexpr/channels
+        mkdir -m 0755 -p /var/lib/nixos
+        touch /var/lib/nixos/did-channel-init
+      fi
     '';
 }
